+++ src/interfaces/IAprovacaoData.tsx
@@ -1,0 +1,13 @@
+
+export interface AprovacaoData {
+  etiqueta: string;
+  material: string;
+  descricaoMaterial: string;
+  op: string;
+  qm: string;
+  qtde: string;
+  status: string;
+  data: string;
+  hora: string;
+  operador: string;
+}
