+++ src/interfaces/IEtiquetaData.tsx
@@ -1,0 +1,18 @@
+export interface EtiquetaData {
+  endereco: string;
+  deposito: string;
+  etiqueta: string;
+  material: string;
+  operador: string;
+  status: string;
+  data: string;
+  hora: string;
+  descricaoMaterial: string;
+  descricaoDeposito: string;
+  descricaoEndereco: string;
+  origem: string;
+  depositoAtual: string;
+  op: string;
+  qm: string;
+  qtde: string;
+}
