+++ src/interfaces/IManualInputProps.tsx
@@ -1,0 +1,8 @@
+export interface ManualInputProps {
+  route: {
+    params?: {
+      userUser?: string;
+      scanType?: string;
+    };
+  };
+}
