+++ src/screens/ManualInput/index.tsx
@@ -9,15 +9,8 @@ import {
 } from "react-native";
 import { useNavigation } from "@react-navigation/native";
 import { RootStackNavigationProp } from "../../types/navigation";
-
-interface ManualInputProps {
-  route: {
-    params?: {
-      userUser?: string;
-      scanType?: string;
-    };
-  };
-}
+import { ManualInputProps } from "../../interfaces/IManualInputProps";
+import styles from "../../styles/styleManualInput";
 
 const ManualInput = ({ route }: ManualInputProps) => {
   const navigation = useNavigation<RootStackNavigationProp>();
@@ -129,56 +122,5 @@ const ManualInput = ({ route }: ManualInputProps) => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-  },
-  content: {
-    flex: 1,
-    padding: 20,
-    justifyContent: "center",
-  },
-  title: {
-    fontSize: 24,
-    marginBottom: 30,
-    color: "#282abd",
-    fontWeight: "bold",
-    textAlign: "center",
-  },
-  input: {
-    width: "100%",
-    borderWidth: 1,
-    borderColor: "#282abd",
-    borderRadius: 8,
-    padding: 15,
-    fontSize: 16,
-    marginBottom: 30,
-    backgroundColor: "#fff",
-  },
-  buttonContainer: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-  },
-  button: {
-    flex: 1,
-    padding: 15,
-    borderRadius: 8,
-    marginHorizontal: 5,
-  },
-  cancelButton: {
-    backgroundColor: "#6c68b5",
-  },
-  confirmButton: {
-    backgroundColor: "#282abd",
-  },
-  buttonText: {
-    color: "white",
-    fontSize: 16,
-    textAlign: "center",
-    fontWeight: "bold",
-  },
-});
 
 export default ManualInput;
