+++ src/screens/Scanner/index.tsx
@@ -11,6 +11,7 @@ import {
 import { CameraView, useCameraPermissions } from "expo-camera";
 import { useNavigation } from "@react-navigation/native";
 import { RootStackNavigationProp } from "../../types/navigation";
+import styles from "../../styles/styleScanner";
 
 const Scanner = ({
   route,
@@ -422,221 +423,5 @@ const Scanner = ({
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  modalContainer: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: "rgba(0, 0, 0, 0.5)",
-    paddingHorizontal: 20,
-  },
-  modalContent: {
-    backgroundColor: "white",
-    borderRadius: 10,
-    padding: 24,
-    width: "90%",
-    maxWidth: 400,
-    marginVertical: "auto",
-    elevation: 5,
-  },
-  buttonContainer: {
-    flexDirection: "row",
-    justifyContent: "space-around",
-    padding: 20,
-    position: "absolute",
-    bottom: 0,
-    left: 0,
-    right: 0,
-    backgroundColor: "rgba(255, 255, 255, 0.8)",
-  },
-  actionButton: {
-    backgroundColor: "#4169E1",
-    padding: 15,
-    borderRadius: 8,
-    width: "45%",
-    elevation: 3,
-  },
-  modalTitle: {
-    fontSize: 18,
-    marginBottom: 20,
-    color: "#282abd",
-    fontWeight: "bold",
-  },
-  input: {
-    width: "100%",
-    borderWidth: 1,
-    borderColor: "#282abd",
-    borderRadius: 4,
-    padding: 8,
-    backgroundColor: "white",
-    color: "#000",
-  },
-  modalButtons: {
-    gap: 12,
-    flexDirection: "column",
-    alignItems: "stretch",
-    width: "100%",
-    marginTop: 20,
-    paddingBottom: 10,
-  },
-  cancelButton: {
-    backgroundColor: "#6c68b5",
-    marginBottom: 8,
-  },
-  editButton: {
-    backgroundColor: "#4169E1",
-    marginBottom: 16,
-    width: "100%",
-  },
-  confirmButton: {
-    backgroundColor: "#282abd",
-    marginTop: 8, // Adicionando margem superior para separaçã
-    marginBottom: 16,
-  },
-  approvalButton: {
-    backgroundColor: "#282abd",
-  },
-  container: {
-    flexDirection: "column",
-    alignItems: "center",
-    justifyContent: "center",
-    flex: 1,
-  },
-  txtHome: {
-    fontSize: 20,
-    marginBottom: 25,
-    color: "#EFB719",
-  },
-  message: {
-    fontSize: 16,
-    textAlign: "center",
-    marginBottom: 20,
-  },
-  cameraContainer: {
-    width: "80%",
-    height: 300,
-    overflow: "hidden",
-    borderRadius: 10,
-    marginBottom: 20,
-  },
-  camera: {
-    flex: 1,
-  },
-  button: {
-    padding: 10,
-    backgroundColor: "#282abd",
-    borderRadius: 8,
-    alignItems: "center",
-  },
-  buttonText: {
-    color: "white",
-    fontSize: 16,
-  },
-  homeButton: {
-    padding: 10,
-    backgroundColor: "#282abd",
-    borderRadius: 8,
-    alignItems: "center",
-    marginTop: 20,
-  },
-  homeButtonText: {
-    color: "white",
-    fontSize: 16,
-  },
-  modalOverlay: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: "rgba(0, 0, 0, 0.5)",
-  },
-  modalContent: {
-    backgroundColor: "#6495ED",
-    padding: 24,
-    borderRadius: 10,
-    width: "90%",
-    maxHeight: "80%",
-    marginVertical: "auto",
-    marginHorizontal: "auto",
-    elevation: 5,
-    marginBottom: 32,
-  },
-  modalRow: {
-    flexDirection: "column",
-    marginBottom: 15,
-    width: "100%",
-  },
-  modalField: {
-    width: "100%",
-    marginBottom: 10,
-  },
-  modalLabel: {
-    fontSize: 16,
-    color: "white",
-    fontWeight: "bold",
-    marginBottom: 6,
-  },
-  modalValue: {
-    fontSize: 18,
-    color: "white",
-    backgroundColor: "rgba(255, 255, 255, 0.3)",
-    padding: 12,
-    borderRadius: 6,
-    marginBottom: 4,
-  },
-  modalValueDesc: {
-    fontSize: 14,
-    color: "white",
-    marginTop: 4,
-    paddingHorizontal: 4,
-  },
-  modalInputContainer: {
-    backgroundColor: "rgba(255, 255, 255, 0.3)",
-    borderRadius: 6,
-    padding: 12,
-  },
-  modalButton: {
-    padding: 15,
-    backgroundColor: "#4169E1",
-    borderRadius: 8,
-    alignItems: "center",
-    marginTop: 16,
-    marginBottom: 16,
-    width: "100%",
-  },
-  modalButtonText: {
-    color: "white",
-    fontSize: 18,
-    fontWeight: "bold",
-  },
-  editButton: {
-    backgroundColor: "#4169E1",
-    marginBottom: 16,
-    width: "100%",
-  },
-  modalText: {
-    fontSize: 16,
-    color: "white",
-    marginBottom: 10,
-  },
-  statusContainer: {
-    padding: 10,
-    backgroundColor: "rgba(255, 255, 255, 0.9)",
-    borderRadius: 8,
-    marginBottom: 10,
-    width: "80%",
-    alignItems: "center",
-  },
-  statusText: {
-    fontSize: 16,
-    color: "#282abd",
-    fontWeight: "bold",
-    marginBottom: 5,
-  },
-  stepIndicator: {
-    fontSize: 14,
-    color: "#666",
-  },
-});
 
 export default Scanner;
